Fix loss of hyperlink functionality.
