+++ WinShell/WinShell/ProcessorCommand.cs
@@ -11,38 +11,8 @@ namespace WinShell
     /// A class representing a generic "Command Processor" command.
     /// Concrete command classes must implement the bottom three methods.
     /// </summary>
-    public abstract class ProcessorCommand : ICommand
+    public abstract class ProcessorCommand
     {
-        /// <summary>
-        /// The main window associated with this command.
-        /// </summary>
-        public MainWindow MainWindow { get; set; }
-
-        /// <summary>
-        /// Returns a value indicating whether the process is in a state that supports executing this command.
-        /// </summary>
-        /// <param name="parameter">The parameter object associated with the command request.</param>
-        /// <returns>A value indicating whether the UI should make this command available to the user via the UI.</returns>
-        public bool CanExecute(object parameter)
-        {
-            return true;
-        }
-
-        /// <summary>
-        /// Executes the command by passing the parameter as a string argument to the CommandProcessor.
-        /// </summary>
-        /// <param name="parameter">The parameter object associated with the command request.</param>
-        public void Execute(object parameter)
-        {
-            MainWindow.ProcessCommand(parameter as string);
-            MainWindow.PresentCommandPrompt();
-        }
-
-        /// <summary>
-        /// An event that listeners can use for dynamically determining when "can execute" status changes.
-        /// </summary>
-        public event EventHandler CanExecuteChanged;
-
         abstract public IEnumerable<string> GetArgs();
         abstract public SingleCommandType GetSingleCommandType();
         abstract public MultiCommandType GetMultiCommandType();
